chore: refactor particles setup in Header component
The code in the Header component has been refactored to improve the setup of the particles animation. Instead of checking for window and particlesRef.current in multiple places, the setupParticles function is now called in the useEffect hook. This function imports the particles.min.js file and then calls the window.particlesJS.load function to load the particles animation. This simplifies the code and ensures that the particles animation is properly set up.